+++ app/assets/javascripts/beyond_canvas/base.js
@@ -3,6 +3,7 @@
  */
 //= require jquery3
 //= require rails-ujs
+//= require beyond_canvas/common_functions
 //= require_self
 
 (function(factory) {